d/requirements.txt
discord.py==2.3.1
pytz==2023.3